Make RemyPacket and Memory tickless doubles. Ref: 2e6461a4e410af82d013afccfe2576d297620c1f Ref: 2e6461a4e410af82d013afccfe2576d297620c1f Start RemyCC only after 3-way handshake.

+++ tcp/remy/packet.hh
@@ -4,9 +4,9 @@
 class RemyPacket
 {
 public:
-  unsigned int tick_sent, tick_received;
+  double tick_sent, tick_received;
 
-  RemyPacket( const unsigned int & s_tick_sent, const unsigned int & s_tick_received )
+  RemyPacket( const double & s_tick_sent, const double & s_tick_received )
     : tick_sent( s_tick_sent ), tick_received( s_tick_received )
   {}
